refactor(app): extract media cache
