+++ src/app/media_cache.rs
@@ -1,0 +1,96 @@
+use std::collections::VecDeque;
+use std::sync::Arc;
+use std::thread;
+
+use ratatui_image::protocol::StatefulProtocol;
+use whatsrust as wr;
+
+use super::media_support::probe_audio_duration;
+use super::{App, AppEvent, AppInput, FileMeta, Metadata};
+
+const IMAGE_CACHE_CAPACITY: usize = 50;
+
+fn touch_order(order: &mut VecDeque<Arc<str>>, path: &Arc<str>) {
+    order.retain(|cached| cached != path);
+    order.push_back(path.clone());
+}
+
+impl App<'_> {
+    pub fn touch_image_cache(&mut self, path: &Arc<str>) {
+        if self.image_cache.contains_key(path) {
+            touch_order(&mut self.image_cache_order, path);
+        }
+    }
+
+    pub(crate) fn cache_file_preview(
+        &mut self,
+        message_id: wr::MessageId,
+        file_path: Arc<str>,
+        preview: StatefulProtocol,
+    ) {
+        if !self.image_cache.contains_key(&file_path)
+            && self.image_cache.len() >= IMAGE_CACHE_CAPACITY
+            && let Some(oldest) = self.image_cache_order.pop_front()
+        {
+            self.image_cache.remove(&oldest);
+            self.mark_evicted_preview_reloadable(&oldest);
+        }
+        self.image_cache.insert(file_path.clone(), preview);
+        self.touch_image_cache(&file_path);
+        self.metadata
+            .insert(message_id.clone(), Metadata::File(FileMeta::Loaded));
+        self.message_height_cache.invalidate(&message_id);
+    }
+
+    fn mark_evicted_preview_reloadable(&mut self, path: &Arc<str>) {
+        for (id, message) in &self.messages {
+            if matches!(&message.message, wr::MessageContent::File(file) if file.path == *path)
+                && matches!(
+                    self.metadata.get(id),
+                    Some(Metadata::File(FileMeta::Loaded))
+                )
+            {
+                self.metadata
+                    .insert(id.clone(), Metadata::File(FileMeta::Downloaded));
+                self.message_height_cache.invalidate(id);
+            }
+        }
+    }
+
+    /// Spawns a background probe for the audio duration of `message_id` once
+    /// its file is on disk. No-op for non-audio messages and for paths that
+    /// already have a cached duration.
+    pub(crate) fn spawn_audio_duration_probe_if_missing(&self, message_id: &wr::MessageId) {
+        let Some(file) = (match self
+            .messages
+            .get(message_id)
+            .map(|message| &message.message)
+        {
+            Some(wr::MessageContent::File(file)) => Some(file.clone()),
+            _ => None,
+        }) else {
+            return;
+        };
+        if !matches!(file.kind, wr::FileKind::Audio)
+            || self.audio_durations.contains_key(file.path.as_ref())
+        {
+            return;
+        }
+        let tx = self.tx.clone();
+        let media_path = self.media_path.to_owned();
+        let message_id = message_id.clone();
+        thread::spawn(move || {
+            let absolute = media_path.join(file.path.as_ref());
+            let duration = probe_audio_duration(&absolute);
+            tx.send(AppInput::App(AppEvent::SetAudioDuration(
+                message_id.clone(),
+                file.path,
+                duration,
+            )))
+            .unwrap();
+        });
+    }
+}
+
+#[cfg(test)]
+mod tests;
